requirements.txt
# Core dependencies
python-telegram-bot==21.0.1
transparent-background==1.3.4

# Image processing dependencies
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4

# Utility dependencies
python-dotenv==1.0.1
asyncio==3.4.3

# Optional GPU acceleration for better performance
# Uncomment if you have CUDA-capable GPU
# torch>=1.7.1
# torchvision>=0.8.2

# For development and testing
pytest==8.3.2
pytest-asyncio==0.24.0
